requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=2.0.23
Flask-Migrate>=4.0.5
sqlalchemy-serializer>=1.4.1
Flask-RESTful>=0.3.10
Flask-CORS>=4.0.0
Flask-JWT-Extended>=4.5.3
Flask-Mail>=0.9.1
PyMySQL>=1.1.0
cloudinary>=1.37.0
sqlalchemy>=2.0.23
cryptography>=41.0.6
setuptools>=69.0.3
gunicorn==21.2.0
flask-restx>=1.3.0
requests>=2.31.0
# passlib==1.7.4
# DPO
certifi==2021.10.8
charset-normalizer==2.0.7
idna==3.3
pydantic==1.8.2
python-dotenv==0.19.2
typing-extensions>=4.2.0
urllib3==1.26.7
xmltodict==0.12.0
africastalking==1.2.8
